+++ kubernetes.tf
@@ -1,8 +1,0 @@
-// See https://registry.terraform.io/providers/hashicorp/google/latest/docs/data-sources/client_config#example-usage-configure-kubernetes-provider-with-oauth2-access-token
-data "google_client_config" "provider" {}
-
-provider "kubernetes" {
-  host                   = "https://${local.cluster_endpoint}"
-  token                  = data.google_client_config.provider.access_token
-  cluster_ca_certificate = base64decode(local.cluster_ca_certificate)
-}